Feat: test for mqtt_client
